8.String to Integer(atoi)
Implement the myAtoi(string s) function, which converts a string to a 32-bit signed integer.

The algorithm for myAtoi(string s) is as follows:

Whitespace: Ignore any leading whitespace (" ").
Signedness: Determine the sign by checking if the next character is '-' or '+', assuming positivity if neither present.
Conversion: Read the integer by skipping leading zeros until a non-digit character is encountered or the end of the string is reached. If no digits were read, then the result is 0.
Rounding: If the integer is out of the 32-bit signed integer range [-231, 231 - 1], then round the integer to remain in the range. Specifically, integers less than -231 should be rounded to -231, and integers greater than 231 - 1 should be rounded to 231 - 1.
Return the integer as the final result.

 

Example 1:

Input: s = "42"

Output: 42

Explanation:

The underlined characters are what is read in and the caret is the current reader position.
Step 1: "42" (no characters read because there is no leading whitespace)
         ^
Step 2: "42" (no characters read because there is neither a '-' nor '+')
         ^
Step 3: "42" ("42" is read in)
           ^

Intuition 🌄
The problem requires converting a string to a 32-bit signed integer.The key challenges are handling edge cases, like leading whitespace, optional signs, and overflow, while making surethat only valid numeric characters added.

Approach 🤔
1.Trim Whitespace using the strip() method to remove leading and trailing whitespace from the string.
- If the string is empty after trimming, return 0.

Handle Signs by checking if the first character is a + or -.
Update the sign variable accordingly (1 for positive, -1 for negative). We just multiply this by result before returning it.

We also have to then move i to the next character so we can skip over the sign.
Process Digits in the string:

Iterate through the string while the current character is a digit (0-9).
Convert the character to its integer value and add it to the result.
Before updating the result, check if adding the next digit would cause an overflow:
If result > (2^31 - 1 - digit) // 10, clamp the result to 2^31 - 1 for positive numbers or -2^31 for negative numbers.
4.return result multiplyed by sign, making sure that it has the correct sign.

Complexity
Time complexity:O(n)
Space complexity:O(1)